refactor(mysql): create util package for extractor
