+++ plugins/extractors/mysql/mysql.go
@@ -16,9 +16,9 @@ import (
 	assetsv1beta1 "github.com/odpf/meteor/models/odpf/assets/v1beta1"
 
 	"github.com/odpf/meteor/plugins"
+	extutil "github.com/odpf/meteor/plugins/extractors/extractorutils"
 	"github.com/odpf/meteor/registry"
 	"github.com/odpf/meteor/utils"
-	"github.com/odpf/salt/log"
 )
 
 //go:embed README.md
@@ -31,31 +31,10 @@ var defaultDBList = []string{
 	"sys",
 }
 
-// Config holds the connection URL for the extractor
-type Config struct {
-	ConnectionURL string `mapstructure:"connection_url" validate:"required"`
-}
-
 var sampleConfig = `connection_url: "admin:pass123@tcp(localhost:3306)/"`
 
-// Extractor manages the extraction of data from MySQL
-type Extractor struct {
-	excludedDbs map[string]bool
-	logger      log.Logger
-	config      Config
-	db          *sql.DB
-	emit        plugins.Emit
-}
-
-// New returns a pointer to an initialized Extractor Object
-func New(logger log.Logger) *Extractor {
-	return &Extractor{
-		logger: logger,
-	}
-}
-
 // Info returns the brief information about the extractor
-func (e *Extractor) Info() plugins.Info {
+func (e *extutil.Extractor) Info() plugins.Info {
 	return plugins.Info{
 		Description:  "Table metadata from MySQL server.",
 		SampleConfig: sampleConfig,
@@ -65,12 +44,12 @@ func (e *Extractor) Info() plugins.Info {
 }
 
 // Validate validates the configuration of the extractor
-func (e *Extractor) Validate(configMap map[string]interface{}) (err error) {
-	return utils.BuildConfig(configMap, &Config{})
+func (e *extutil.Extractor) Validate(configMap map[string]interface{}) (err error) {
+	return utils.BuildConfig(configMap, &extutil.Config{})
 }
 
 // Init initializes the extractor
-func (e *Extractor) Init(ctx context.Context, configMap map[string]interface{}) (err error) {
+func (e *extutil.Extractor) Init(ctx context.Context, configMap map[string]interface{}) (err error) {
 	if err = utils.BuildConfig(configMap, &e.config); err != nil {
 		return plugins.InvalidConfigError{}
 	}
@@ -88,7 +67,7 @@ func (e *Extractor) Init(ctx context.Context, configMap map[string]interface{}) 
 
 // Extract extracts the data from the MySQL server
 // and collected through the emitter
-func (e *Extractor) Extract(ctx context.Context, emit plugins.Emit) (err error) {
+func (e *extutil.Extractor) Extract(ctx context.Context, emit plugins.Emit) (err error) {
 	defer e.db.Close()
 	e.emit = emit
 
@@ -113,7 +92,7 @@ func (e *Extractor) Extract(ctx context.Context, emit plugins.Emit) (err error) 
 }
 
 // Extract tables from a given database
-func (e *Extractor) extractTables(database string) (err error) {
+func (e *extutil.Extractor) extractTables(database string) (err error) {
 	// skip if database is default
 	if e.isExcludedDB(database) {
 		return
@@ -145,7 +124,7 @@ func (e *Extractor) extractTables(database string) (err error) {
 }
 
 // processTable builds and push table to emitter
-func (e *Extractor) processTable(database string, tableName string) (err error) {
+func (e *extutil.Extractor) processTable(database string, tableName string) (err error) {
 	var columns []*facetsv1beta1.Column
 	if columns, err = e.extractColumns(tableName); err != nil {
 		return errors.Wrap(err, "failed to extract columns")
@@ -166,7 +145,7 @@ func (e *Extractor) processTable(database string, tableName string) (err error) 
 }
 
 // Extract columns from a given table
-func (e *Extractor) extractColumns(tableName string) (columns []*facetsv1beta1.Column, err error) {
+func (e *extutil.Extractor) extractColumns(tableName string) (columns []*facetsv1beta1.Column, err error) {
 	query := `SELECT COLUMN_NAME,column_comment,DATA_TYPE,
 				IS_NULLABLE,IFNULL(CHARACTER_MAXIMUM_LENGTH,0)
 				FROM information_schema.columns
@@ -199,7 +178,7 @@ func (e *Extractor) extractColumns(tableName string) (columns []*facetsv1beta1.C
 }
 
 // buildExcludedDBs builds the list of excluded databases
-func (e *Extractor) buildExcludedDBs() {
+func (e *extutil.Extractor) buildExcludedDBs() {
 	excludedMap := make(map[string]bool)
 	for _, db := range defaultDBList {
 		excludedMap[db] = true
@@ -209,20 +188,20 @@ func (e *Extractor) buildExcludedDBs() {
 }
 
 // isExcludedDB checks if the given db is in the list of excluded databases
-func (e *Extractor) isExcludedDB(database string) bool {
+func (e *extutil.Extractor) isExcludedDB(database string) bool {
 	_, ok := e.excludedDbs[database]
 	return ok
 }
 
 // isNullable checks if the given string is null or not
-func (e *Extractor) isNullable(value string) bool {
+func (e *extutil.Extractor) isNullable(value string) bool {
 	return value == "YES"
 }
 
 // init register the extractor to the catalog
 func init() {
 	if err := registry.Extractors.Register("mysql", func() plugins.Extractor {
-		return New(plugins.GetLog())
+		return extutil.New(plugins.GetLog())
 	}); err != nil {
 		panic(err)
 	}
